requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.6
psycopg2-binary==2.8.6
sqlparse==0.5.3
whitenoise==6.9.0
